Sample Updates for Consistency
I'm attempting to show that the library has a consistent set of
functions for each Card type (as much as possible).  But the samples
weren't totally consistent in function naming conventions, which makes
it hard to compare differences, so that's fixed.  Again, minor.  Just
trying to improve the developer experience.