Fix a potential issue with IS23SC4442 cards
I realized that the buffer handling that I implemented for IS23SC4442
addressable read support may be an issue if you swapped cards but didn't
write to the new card before performing a read (i.e. the ReadBuffer
would still have data from the previous card in it).   So, I've created
a private callback that gets called so that when a new card is inserted,
the ReadBuffer for the IS23SC4442 card is refreshed so the read values
would be correct for the newly inserted card.   Note:  The private
callback is also called for the other cards, but those cards don't need
to be refreshed, since the addressable read is directly from the cards,
not from a read buffer handled by the library.